+++ tests/mocks/ExecutionEnvironment.ts
@@ -1,8 +1,0 @@
-const ExecutionEnvironment = {
-  canUseDOM: true,
-  canUseEventListeners: true,
-  canUseIntersectionObserver: true,
-  canUseViewport: true,
-}
-
-export default ExecutionEnvironment